chore: split routes, begin debugging
