+++ examples/placegoose/src/routes/index.ts
@@ -1,0 +1,3 @@
+export { default as gaggles } from './gaggles'
+export { default as geese } from './geese'
+export { default as honks } from './honks'
