+++ examples/placegoose/src/types.ts
@@ -1,0 +1,3 @@
+export type Bindings = {
+    DB: D1Database;
+  };
